enable auth for all api endpoints, add userinfo function
the userinfo call returns info to a specific Supervisor / Admin-User object given its username and thus allows the AuthService of the frontend to retrieve relevant data in a call after the authentication succeeded. This could be implemented as a subclass of TokenAuth in order to return that information together with the token instead of seperately.