Fix Attendance generation in models save() method
Same as previous commit, this fixes the generation of attendance upon generation of a specific date such that only subscription active at the given date are considered. This is somewhat a duplicate of what happens in the serializer, maybe it can be removed, since whenever the model is queried the attendance is updated.