Fix Subscription conflict problems when subscription is updated.
Problem was, that when a subscription is edited / updated, it naturally discovers conflicts. Since I found no way to check within the serializer, whether the object is added or edited, I implemented a new Serializer for that purpose, which respects the fact that one collision is ok.

TODO: Implement some sort of DB check or check for ID